+++ rocketmq-20220801/alibabacloud_rocketmq20220801/models/_create_instance_request.py
@@ -302,6 +302,7 @@ class CreateInstanceRequestProductInfo(DaraModel):
         self,
         auto_scaling: bool = None,
         capacity_type: str = None,
+        dr_replication_mode: str = None,
         message_retention_time: int = None,
         msg_process_spec: str = None,
         provisioned_capacity: int = None,
@@ -322,6 +323,7 @@ class CreateInstanceRequestProductInfo(DaraModel):
         # >  The elastic TPS feature is supported only by instances of specific editions. For more information, see [Instance editions](https://help.aliyun.com/document_detail/444715.html).
         self.auto_scaling = auto_scaling
         self.capacity_type = capacity_type
+        self.dr_replication_mode = dr_replication_mode
         # The retention period of messages. Unit: hours.
         # 
         # For information about the valid values of this parameter, see the "Limits on resource quotas" section of the [Limits](https://help.aliyun.com/document_detail/440347.html) topic.
@@ -357,6 +359,9 @@ class CreateInstanceRequestProductInfo(DaraModel):
         if self.capacity_type is not None:
             result['capacityType'] = self.capacity_type
 
+        if self.dr_replication_mode is not None:
+            result['drReplicationMode'] = self.dr_replication_mode
+
         if self.message_retention_time is not None:
             result['messageRetentionTime'] = self.message_retention_time
 
@@ -387,6 +392,9 @@ class CreateInstanceRequestProductInfo(DaraModel):
 
         if m.get('capacityType') is not None:
             self.capacity_type = m.get('capacityType')
+
+        if m.get('drReplicationMode') is not None:
+            self.dr_replication_mode = m.get('drReplicationMode')
 
         if m.get('messageRetentionTime') is not None:
             self.message_retention_time = m.get('messageRetentionTime')